Removing compilation warnings on the Alpha
git-svn-id: https://stealth.svn.sourceforge.net/svnroot/yodl/trunk@125 7de777c3-3b1c-0410-83c5-80958bd9f4b1

+++ yodl/src/yodlpost/constructverb.c
@@ -14,5 +14,5 @@
 HashItem *construct_verb(char const *key, char *rest)
 {
     return hashitem_construct(VOIDPTR, "",
-                    (void *)(*(rest + 1) == 'n'), root_nop);
+                    (void *)(size_t)(*(rest + 1) == 'n'), root_nop);
 }
